All Changes to api response and tests to Comments

+++ StockPricePrediction/Layers/ServiceLayer/Comment/CommentService.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using RepositoryLayer;
 using DomainLayer;
+using ServiceLayer.Models;
 
 
 namespace ServiceLayer
@@ -9,7 +11,7 @@ namespace ServiceLayer
     public class CommentService : ICommentService
     {
         private readonly ICommentRepository _repository;
-        private IStockRepository _stockRepository;
+        private readonly IStockRepository _stockRepository;
 
         public CommentService(ICommentRepository repository, IStockRepository stockRepository)
         {
@@ -17,48 +19,79 @@ namespace ServiceLayer
             _stockRepository = stockRepository;
         }
 
-        public IEnumerable<Comment> GetAllComments(string symbol)
+        public ApiResponse<IEnumerable<Comment>> GetAllComments(string symbol)
         {
             var stock = _stockRepository.GetBySymbol(symbol);
-            return _repository.GetAll(stock);
+            return stock == null
+                ? ApiResponse<IEnumerable<Comment>>.Fail("Stocks Empty")
+                : ApiResponse<IEnumerable<Comment>>.Success(_repository.GetAll(stock));
         }
 
-        public Comment GetComment(int id)
+        public ApiResponse<Comment> GetComment(int id)
         {
-            return _repository.Get(id);
+            if (id < 0)
+                return ApiResponse<Comment>.Fail("Invalid id");
+            var comment = _repository.Get(id);
+            return comment != null
+                ? ApiResponse<Comment>.Success(comment)
+                : ApiResponse<Comment>.Fail("Comment does not exist");
         }
 
 
-        public int InsertComment(string author, string message, string stockSymbol, DateTime date)
+        public ApiResponse<int> InsertComment(string author, string message, string stockSymbol, DateTime date)
         {
-            var comment = new Comment(author, message);
-            comment.CreationDate = date;
+            var comment = new Comment(author, message)
+            {
+                CreationDate = date
+            };
+            if (author == null || message == null || stockSymbol == null)
+            {
+                return ApiResponse<int>.Fail("Invalid comment");
+            }
+
             _repository.Insert(comment);
-            return _stockRepository.AddComment(comment, stockSymbol);
+            return ApiResponse<int>.Success(_stockRepository.AddComment(comment, stockSymbol));
         }
 
-        public void UpdateComment(Comment comment)
+        public ApiResponse<bool> UpdateComment(Comment comment)
         {
+            if (comment == null)
+                return ApiResponse<bool>.Fail("Null comment");
             _repository.Update(comment);
+            return ApiResponse<bool>.Success(true);
         }
 
-        public void DeleteComment(int id)
+        public ApiResponse<bool> DeleteComment(int id)
         {
-            var comment = GetComment(id);
+            if (id < 0)
+            {
+                return ApiResponse<bool>.Fail("Invalid id");
+            }
+
+            var comment = GetComment(id).Data;
             _repository.Remove(comment);
             _repository.SaveChanges();
+            return ApiResponse<bool>.Success(true);
         }
 
-        public void Upvote(int id)
+        public ApiResponse<bool> Upvote(int id)
         {
             var comment = GetComment(id);
-            _repository.UpVote(comment);
+            return comment == null
+                ? ApiResponse<bool>.Fail("Null Comment cannot be upvoted")
+                : ApiResponse<bool>.Success(true);
         }
 
-        public void Downvote(int id)
+        public ApiResponse<bool> Downvote(int id)
         {
             var comment = GetComment(id);
-            _repository.DownVote(comment);
+            if (comment.Data == null)
+            {
+                return ApiResponse<bool>.Fail("Comment doesn't exist");
+            }
+
+            _repository.DownVote(comment.Data);
+            return ApiResponse<bool>.Success(true);
         }
     }
 }
